Move test utilities into testutil
