+++ java/clients/src/test/java/sleeper/clients/admin/AdminCommonPromptsTest.java
@@ -20,8 +20,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import sleeper.clients.testutil.TestConsoleInput;
 import sleeper.clients.testutil.ToStringPrintStream;
-import sleeper.clients.util.console.TestConsoleInput;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
 
@@ -33,7 +33,7 @@ import static sleeper.clients.admin.AdminCommonPrompts.confirmReturnToMainScreen
 import static sleeper.clients.admin.AdminCommonPrompts.tryLoadInstanceProperties;
 import static sleeper.clients.admin.AdminCommonPrompts.tryLoadTableProperties;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.PROMPT_RETURN_TO_MAIN;
-import static sleeper.clients.util.console.TestConsoleInput.CONFIRM_PROMPT;
+import static sleeper.clients.testutil.TestConsoleInput.CONFIRM_PROMPT;
 
 public class AdminCommonPromptsTest {
     private final ToStringPrintStream out = new ToStringPrintStream();
