+++ java/clients/src/test/java/sleeper/clients/admin/PartitionsStatusReportScreenTest.java
@@ -36,8 +36,8 @@ import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.PROMPT_
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.RETURN_TO_MAIN_SCREEN_OPTION;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.TABLE_SELECT_SCREEN;
 import static sleeper.clients.status.report.partitions.PartitionStatusReportTestHelper.createPartitionsBuilder;
+import static sleeper.clients.testutil.TestConsoleInput.CONFIRM_PROMPT;
 import static sleeper.clients.util.console.ConsoleOutput.CLEAR_CONSOLE;
-import static sleeper.clients.util.console.TestConsoleInput.CONFIRM_PROMPT;
 
 class PartitionsStatusReportScreenTest extends AdminClientMockStoreBase {
     @Test
