+++ java/clients/src/test/java/sleeper/clients/testutil/RunCommandTestHelper.java
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.util;
+package sleeper.clients.testutil;
+
+import sleeper.clients.util.RunCommand;
 
 import java.io.IOException;
 import java.util.concurrent.atomic.AtomicReference;
